move the carrier Scoring function to the carrier package and make it public.
